include/define.h, include/libproto.h, common/item.c, common/readable.c,   common/loader.l: Make safe_strcat() (defined in include/define.h for some   odd reason) use size_t instead of int for buffer size, since this includes   changing a pointer to int into a pointer to size_t update all calls to   safe_strcat() to use size_t for that.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@11316 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ include/define.h
@@ -849,7 +849,7 @@ typedef unsigned char MoveType;
  * @param maxlen
  * maximum length of dest buffer.
  */
-static inline void safe_strcat(char *dest, const char *orig, int *curlen, int maxlen) {
+static inline void safe_strcat(char *dest, const char *orig, size_t *curlen, size_t maxlen) {
     if (*curlen == (maxlen-1))
         return;
     strncpy(dest+*curlen, orig, maxlen-*curlen-1);
